+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 The following is a curated list of changes in the Enact docs-utils module, newest changes on the top.
 
-## Unreleased
+## [unreleased]
 
+- Updated `jsonata` dependency to version 2.x and refactored code to use async functions.
 - Updated `mkdirp` version to 3.0.0 and applied a new API.
 
 ## [0.4.2] - 2023-04-11